as.rle {statnet.common}	R Documentation

Coerce to rle if not already an rle object.

Description

Coerce to rle if not already an rle object.

Usage

as.rle(x)

## S3 method for class 'rle'
as.rle(x)

## Default S3 method:
as.rle(x)

Arguments

x
the object to be coerced.

[Package statnet.common version 4.1.2 Index]