compress.data.frame {statnet.common}	R Documentation

"Compress" a data frame.

Description

compress.data.frame "compresses" a data frame, returning unique rows and a tally of the number of times each row is repeated, as well as a permutation vector that can reconstruct the original data frame. decompress.data.frame reconstructs the original data frame.

Usage

compress.data.frame(x)

decompress.data.frame(x)

Arguments

x
For compress.data.frame a data.frame to be compressed. For decompress.data.frame a list as returned by compress.data.frame.

Value

For compress.data.frame, a list with three elements:

rows
Unique rows of x

frequencies
A vector of the same length as the number or rows, giving the number of times the corresponding row is repeated

ordering
A vector such that if c is the compressed data frame, c$rows[c$ordering,,drop=FALSE] equals the original data frame, except for row names

rownames
Row names of x

For decompress.data.frame, the original data frame.

See Also

data.frame

Examples

(x <- data.frame(V1=sample.int(3,30,replace=TRUE),
                 V2=sample.int(2,30,replace=TRUE),
                 V3=sample.int(4,30,replace=TRUE)))

(c <- compress.data.frame(x))

stopifnot(all(decompress.data.frame(c)==x))

[Package statnet.common version 4.1.2 Index]