control.remap {statnet.common}	R Documentation

Overwrite control parameters of one configuration with another.

Description

Given a control.list, and two prefixes, from and to, overwrite the elements starting with to with the corresponding elements starting with from.

Usage

control.remap(control, from, to)

Arguments

control
An object of class control.list.

from
Prefix of the source of control parameters.

to
Prefix of the destination of control parameters.

Value

An control.list object.

Author(s)

Pavel N. Krivitsky

See Also

print.control.list

Examples

(l <- set.control.class("test", list(a.x=1, a.y=2)))
control.remap(l, "a", "b")

[Package statnet.common version 4.1.2 Index]