diff.control.list {statnet.common}	R Documentation

Identify and the differences between two control lists.

Description

Identify and the differences between two control lists.

Usage

## S3 method for class 'control.list'
diff(x, y = eval(call(class(x)[[1L]])), ignore.environment = TRUE, ...)

## S3 method for class 'diff.control.list'
print(x, ..., indent = "")

Arguments

x
a control.list

y
a reference control.list; defaults to the default settings for x.

ignore.environment
whether environment for environment-bearing parameters (such as formulas and functions) should be considered when comparing.

...
Additional arguments to methods.

indent
an argument for recursive calls, to facilitate indentation of nested lists.

Value

An object of class diff.control.list: a named list with an element for each non-identical setting. The element is either itself a diff.control.list (if the setting is a control list) or a named list with elements x and y, containing x's and y's values of the parameter for that setting.

Methods (by generic)

print(diff.control.list): A print method.

[Package statnet.common version 4.13.0 Index]