envir {statnet.common}	R Documentation

A generic for querying and setting an object's environment

Description

environment() and environment<-() are not generics, so it is not possible to dispatch based on the class of the object affected.

Usage

envir(object)

envir(object) <- value

Arguments

object
object whose environment is to be queried or set

value
typically an environment, but could be any RHS supported by the method

Details

When no method is available, these generics fall back to the environment() and environment<-() functions.

[Package statnet.common version 4.13.0 Index]