is.SPD {statnet.common}	R Documentation

Test if the object is a matrix that is symmetric and positive definite

Description

Test if the object is a matrix that is symmetric and positive definite

Usage

is.SPD(x, tol = .Machine$double.eps)

Arguments

x
the object to be tested.

tol
the tolerance for the reciprocal condition number.

[Package statnet.common version 4.13.0 Index]