message_print {statnet.common}	R Documentation

print objects to the message output.

Description

A thin wrapper around print that captures its output and prints it as a message, usually to STDERR.

Usage

message_print(..., messageArgs = NULL)

Arguments

...
arguments to print.

messageArgs
a list of arguments to be passed directly to message.

Examples

cat(1:5)

print(1:5)
message_print(1:5) # Looks the same (though may be in a different color on some frontends).

suppressMessages(print(1:5)) # Still prints
suppressMessages(message_print(1:5)) # Silenced

[Package statnet.common version 4.1.2 Index]