opttest {statnet.common}	R Documentation

Optionally test code depending on environment variable.

Description

A convenience wrapper to run code based on whether an environment variable is defined.

Usage

opttest(expr, testname = NULL, testvar = "ENABLE_statnet_TESTS",
  yesvals = c("y", "yes", "t", "true", "1"), lowercase = TRUE)

Arguments

expr
An expression to be evaluated only if testvar is set to a non-empty value.

testname
Optional name of the test. If given, and the test is skipped, will print a message to that end, including the name of the test, and instructions on how to enable it.

testvar
Environment variable name. If set to one of the yesvals, expr is run. Otherwise, an optional message is printed.

yesvals
A character vector of strings considered affirmative values for testvar.

lowercase
Whether to convert the value of testvar to lower case before comparing it to yesvals.

[Package statnet.common version 4.1.2 Index]