print.control.list {statnet.common}	R Documentation

Pretty print the control list

Description

This function prints the control list, including what it can control and the elements.

Usage

## S3 method for class 'control.list'
print(x, ..., indent = "")

Arguments

x
A list generated by a control.* function.

...
Additional argument to print methods for individual settings.

indent
an argument for recursive calls, to facilitate indentation of nested lists.

See Also

check.control.class, set.control.class

[Package statnet.common version 4.13.0 Index]