set.control.class {statnet.common}	R Documentation

Set the class of the control list

Description

This function sets the class of the control list, with the default being the name of the calling function.

Usage

set.control.class(
  myname = as.character(ult(sys.calls(), 2)[[1L]]),
  control = get("control", pos = parent.frame())
)

Arguments

myname
Name of the class to set.

control
Control list. Defaults to the control variable in the calling function.

Value

The control list with class set.

Note

In earlier versions, OKnames and myname were autodetected. This capability has been deprecated and results in a warning issued once per session. They now need to be set explicitly.

See Also

check.control.class(), print.control.list()

[Package statnet.common version 4.13.0 Index]