sign<- {statnet.common}	R Documentation

A generic for setting the sign of an object

Description

A generic for setting the sign of an object

Usage

sign(x) <- value

Arguments

x
object whose sign is to be set

value
a numeric vector specifying the sign

[Package statnet.common version 4.13.0 Index]