skip_if_not_checking {statnet.common}	R Documentation

Skip a test if not called via R CMD check

Description

A testthat predicate to skip tests if not run as a part of a package check.

Usage

skip_if_not_checking()

[Package statnet.common version 4.13.0 Index]