snctrl {statnet.common}	R Documentation

Statnet Control

Description

A utility to facilitate argument completion of control lists.

Usage

snctrl(...)

Arguments

...
The parameter list is updated dynamically as packages are loaded and unloaded. Their current list is given below.

Details

In and of itself, snctrl copies its named arguments into a list. However, its argument list is updated dynamically as packages are loaded, as are those of its reexports from other packages. This is done using an API provided by helper functions. (See API?snctrl.)

Currently recognised control parameters

This list is updated as packages are loaded and unloaded.

Note

You may see messages along the lines of

The following object is masked from 'package:PKG':
snctrl

when loading packages. They are benign.

[Package statnet.common version 4.13.0 Index]