statnet.cite {statnet.common}	R Documentation

CITATION file utilities for Statnet packages (DEPRECATED)

Description

These functions automate citation generation for Statnet Project packages. They no longer appear to work with CRAN and are thus deprecated.

Usage

statnet.cite.head(pkg)

statnet.cite.foot(pkg)

statnet.cite.pkg(pkg)

Arguments

pkg
Name of the package whose citation is being generated.

Value

For statnet.cite.head and statnet.cite.foot, an object of type citationHeader and citationFooter, respectively, understood by the citation function, with package name substituted into the template.

For statnet.cite.pkg, an object of class bibentry containing a 'software manual' citation for the package constructed from the current version and author information in the DESCRIPTION and a template.

See Also

citation, citHeader, citFooter, bibentry

Examples

## Not run: 
statnet.cite.head("statnet.common")

statnet.cite.pkg("statnet.common")

statnet.cite.foot("statnet.common")

## End(Not run)

[Package statnet.common version 4.1.2 Index]