statnetStartupMessage {statnet.common}	R Documentation

Construct a "standard" startup message to be printed when the package is loaded.

Description

This function uses information returned by packageDescription() to construct a standard package startup message according to the policy of the Statnet Project.

Usage

statnetStartupMessage(pkgname, friends = c(), nofriends = c())

Arguments

pkgname
Name of the package whose information is used.

friends, nofriends
No longer used.

Value

A string containing the startup message, to be passed to the packageStartupMessage() call or NULL, if policy prescribes printing default startup message. (Thus, if statnetStartupMessage() returns NULL, the calling package should not call packageStartupMessage() at all.)

Note

Earlier versions of this function printed a more expansive message. This may change again as the Statnet Project policy evolves.

See Also

packageDescription(), packageStartupMessage()

Examples

## Not run: 
.onAttach <- function(lib, pkg){
  sm <- statnetStartupMessage("ergm")
  if(!is.null(sm)) packageStartupMessage(sm)
}

## End(Not run)

[Package statnet.common version 4.13.0 Index]