term_list {statnet.common}	R Documentation

A helper class for list of terms in an formula

Description

Typically generated by list_rhs.formula(), it contains, in addition to a list of call() or similar objects information about the sign of the term and the environment of the formula from which the term has been extracted, accessible and modifiable via sign() and envir() generics. Indexing and concatenation methods preserve these.

Usage

term_list(x, sign = +1L, env = NULL)

as.term_list(x, ...)

## S3 method for class 'term_list'
as.term_list(x, ...)

## Default S3 method:
as.term_list(x, sign = +1L, env = NULL, ...)

## S3 method for class 'term_list'
c(x, ...)

## S3 method for class 'term_list'
x[i, ...]

## S3 method for class 'term_list'
print(x, ...)

## S3 method for class 'term_list'
sign(x)

## S3 replacement method for class 'term_list'
sign(x) <- value

## S3 method for class 'term_list'
envir(object)

## S3 replacement method for class 'term_list'
envir(object) <- value

Arguments

x, object
a list of terms or a term; a term_list

sign
a vector specifying the signs associated with each term (-1 and +1)

env
a list specifying the environments, or NULL

...
additional arguments to methods

i
list index

value
RHS; see method documentation

Methods (by generic)

sign(term_list): An integer vector giving the signs of each term in the list.

sign(term_list) <- value: Update the signs of the terms; value is recycled to the length of the list.

envir(term_list): A list with an element for each term in the list, giving its environment.

envir(term_list) <- value: Update the environments of the terms; value can be an environment or a list of environments, recycled to the length of the term list.

See Also

list_rhs.formula(), list_summands.call()

Examples

e1 <- new.env()
f1 <- a~b+c
environment(f1) <- e1
f2 <- ~-NULL+1

(l1 <- list_rhs.formula(f1))
(l2 <- list_rhs.formula(f2))

(l <- c(l1,l2))

(l <- c(l2[1], l1[2], l1[1], l1[1], l2[2]))
sign(l)[3] <- -1L

[Package statnet.common version 4.13.0 Index]