all_identical {statnet.common}	R Documentation

Test if all items in a vector or a list are identical.

Description

Test if all items in a vector or a list are identical.

Usage

all_identical(x)

Arguments

x
a vector or a list

Value

TRUE if all elements of x are identical to each other.

See Also

identical

Examples

stopifnot(!all_identical(1:3))

stopifnot(all_identical(list("a", "a", "a")))

[Package statnet.common version 4.1.2 Index]