check.control.class {statnet.common}	R Documentation

Check if the class of the control list is one of those that can be used by the calling function

Description

This function can be called to check that the control list passed is appropriate for the function to be controlled. It does so by looking up the class of the control argument (defaulting to the control variable in the calling function) and checking if it matches a list of acceptable classes.

Usage

check.control.class(OKnames = {     sc <- sys.calls()    
  as.character(sc[[length(sc) - 1]][[1]]) }, myname = {     sc <- sys.calls() 
     as.character(sc[[length(sc) - 1]][[1]]) }, control = get("control", pos =
  parent.frame()))

Arguments

OKnames
List of control function names which are acceptable.

myname
Name of the calling function (used in the error message).

control
The control list. Defaults to the control variable in the calling function.

Note

In earlier versions, OKnames and myname were autodetected. This capability has been deprecated and results in a warning issued once per session. They now need to be set explicitly.

See Also

set.control.class, print.control.list

[Package statnet.common version 4.1.2 Index]